# backend/openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: support@viewpredict.ai
    name: ViewPredict AI Support
    url: https://viewpredict.ai/support
  description: "Complete API for YouTube view count prediction platform with user\
    \ authentication, \nprediction management, analytics, and channel integration.\n"
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: ViewPredict AI - YouTube View Count Prediction API
  version: 1.0.0
servers:
- description: Production server
  url: https://api.viewpredict.ai/v1
- description: Staging server
  url: https://staging-api.viewpredict.ai/v1
- description: Local development server
  url: http://localhost:3000/v1
security:
- BearerAuth: []
...
paths:
...
  /channels/callback:
    get:
      description: Handle OAuth callback from YouTube
      parameters:
      - example: 4/0AY0e-g7...
        explode: true
        in: query
        name: code
        required: true
        schema:
          type: string
        style: form
      - example: abc123xyz
        explode: true
        in: query
        name: state
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_callback_get_200_response"
          description: Channel connected successfully
        "400":
          $ref: "#/components/responses/BadRequest"
      security: []
      summary: OAuth callback
      tags:
      - Channels
  /channels/me:
    delete:
      description: Disconnect YouTube channel from account
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_delete_200_response"
          description: Channel disconnected successfully
      summary: Disconnect channel
      tags:
      - Channels
    get:
      description: Retrieve connected YouTube channel information
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_callback_get_200_response"
          description: Channel information retrieved
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_get_404_response"
          description: No channel connected
      summary: Get connected channel
      tags:
      - Channels
...
  /channels/me/videos:
    get:
      description: Retrieve videos from connected YouTube channel
      parameters:
      - explode: true
        in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 50
          minimum: 1
          type: integer
        style: form
      - explode: true
        in: query
        name: sortBy
        required: false
        schema:
          default: publishedAt
          enum:
          - publishedAt
          - viewCount
          - likeCount
          type: string
        style: form
      - explode: true
        in: query
        name: order
        required: false
        schema:
          default: desc
          enum:
          - asc
          - desc
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_videos_get_200_response"
          description: Videos retrieved successfully
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_videos_get_404_response"
          description: No channel connected
      summary: Get channel videos
      tags:
      - Channels
  /channels/me/analytics:
    get:
      description: Retrieve analytics data from connected YouTube channel
      parameters:
      - explode: true
        in: query
        name: period
        required: false
        schema:
          default: 30d
          enum:
          - 7d
          - 30d
          - 3m
          - 1y
          type: string
        style: form
      - example: "views,subscribers"
        explode: false
        in: query
        name: metrics
        required: false
        schema:
          items:
            enum:
            - views
            - subscribers
            - estimatedMinutesWatched
            - averageViewDuration
            type: string
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_analytics_get_200_response"
          description: Channel analytics retrieved
      summary: Get channel analytics
      tags:
      - Channels
  /channels/me/settings:
    get:
      description: Retrieve channel-specific settings
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_settings_get_200_response"
          description: Settings retrieved successfully
      summary: Get channel settings
      tags:
      - Channels
    put:
      description: Update channel-specific settings
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelSettings"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_channels_me_settings_get_200_response"
          description: Settings updated successfully
      summary: Update channel settings
      tags:
      - Channels
...
components:
...
  responses:
    BadRequest:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/inline_object"
      description: Bad request - validation error
...
  schemas:
...
    ConnectedChannel:
      example:
        videoCount: 342
        thumbnail: https://yt3.ggpht.com/abc123
        connectedAt: 2025-09-01T10:00:00Z
        channelUrl: https://youtube.com/@techtutorialshub
        lastSyncAt: 2025-10-03T08:00:00Z
        subscriberCount: 250000
        channelName: Tech Tutorials Hub
        viewCount: 15000000
        channelId: UCabc123xyz
        syncStatus: active
      properties:
        channelId:
          example: UCabc123xyz
          title: channelId
          type: string
        channelName:
          example: Tech Tutorials Hub
          title: channelName
          type: string
        channelUrl:
          example: https://youtube.com/@techtutorialshub
          format: uri
          title: channelUrl
          type: string
        thumbnail:
          example: https://yt3.ggpht.com/abc123
          format: uri
          title: thumbnail
          type: string
        subscriberCount:
          example: 250000
          title: subscriberCount
          type: integer
        videoCount:
          example: 342
          title: videoCount
          type: integer
        viewCount:
          example: 15000000
          title: viewCount
          type: integer
        connectedAt:
          example: 2025-09-01T10:00:00Z
          format: date-time
          title: connectedAt
          type: string
        lastSyncAt:
          example: 2025-10-03T08:00:00Z
          format: date-time
          title: lastSyncAt
          type: string
        syncStatus:
          enum:
          - active
          - syncing
          - error
          example: active
          title: syncStatus
          type: string
      title: ConnectedChannel
      type: object
    ChannelVideo:
      example:
        duration: 720
        thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg
        publishedAt: 2025-09-15T14:00:00Z
        hasPrediction: true
        description: Learn something amazing in this video
        videoId: dQw4w9WgXcQ
        likeCount: 8500
        viewCount: 125000
        title: Amazing Tutorial Video
        predictionId: pred_abc123
        commentCount: 342
      properties:
        videoId:
          example: dQw4w9WgXcQ
          title: videoId
          type: string
        title:
          example: Amazing Tutorial Video
          title: title
          type: string
        description:
          example: Learn something amazing in this video
          title: description
          type: string
        thumbnail:
          example: https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg
          format: uri
          title: thumbnail
          type: string
        publishedAt:
          example: 2025-09-15T14:00:00Z
          format: date-time
          title: publishedAt
          type: string
        duration:
          description: Duration in seconds
          example: 720
          title: duration
          type: integer
        viewCount:
          example: 125000
          title: viewCount
          type: integer
        likeCount:
          example: 8500
          title: likeCount
          type: integer
        commentCount:
          example: 342
          title: commentCount
          type: integer
        hasPrediction:
          example: true
          title: hasPrediction
          type: boolean
        predictionId:
          example: pred_abc123
          nullable: true
          title: predictionId
          type: string
      title: ChannelVideo
      type: object
    ChannelAnalytics:
      example:
        period: 30d
        timeline:
        - date: 2025-10-01
          subscribers: 85
          estimatedMinutesWatched: 3500
          views: 15000
        - date: 2025-10-01
          subscribers: 85
          estimatedMinutesWatched: 3500
          views: 15000
        metrics:
          subscribers: 2500
          estimatedMinutesWatched: 125000
          averageViewDuration: 180
          views: 500000
        topVideos:
        - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          videoId: dQw4w9WgXcQ
          title: Top Performing Video
          views: 50000
        - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          videoId: dQw4w9WgXcQ
          title: Top Performing Video
          views: 50000
        - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          videoId: dQw4w9WgXcQ
          title: Top Performing Video
          views: 50000
        - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          videoId: dQw4w9WgXcQ
          title: Top Performing Video
          views: 50000
        - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          videoId: dQw4w9WgXcQ
          title: Top Performing Video
          views: 50000
        demographics:
          gender:
            other: 1.7
            female: 35.8
            male: 62.5
          topCountries:
          - country: United States
            countryCode: US
            percentage: 45.2
          - country: United States
            countryCode: US
            percentage: 45.2
          ageGroups:
          - ageRange: 18-24
            percentage: 35.5
          - ageRange: 18-24
            percentage: 35.5
      properties:
        period:
          example: 30d
          title: period
          type: string
        metrics:
          $ref: "#/components/schemas/ChannelAnalytics_metrics"
        timeline:
          items:
            $ref: "#/components/schemas/ChannelAnalytics_timeline_inner"
          title: timeline
          type: array
        topVideos:
          items:
            $ref: "#/components/schemas/ChannelAnalytics_topVideos_inner"
          maxItems: 10
          title: topVideos
          type: array
        demographics:
          $ref: "#/components/schemas/ChannelAnalytics_demographics"
      title: ChannelAnalytics
      type: object
    ChannelSettings:
      example:
        syncFrequency: daily
        autoCreatePrediction: false
        autoTracking: true
        notifyOnNewVideo: true
      properties:
        autoTracking:
          description: Automatically track predictions for new videos
          example: true
          title: autoTracking
          type: boolean
        syncFrequency:
          description: How often to sync channel data
          enum:
          - hourly
          - daily
          - weekly
          example: daily
          title: syncFrequency
          type: string
        notifyOnNewVideo:
          description: Send notification when new video is detected
          example: true
          title: notifyOnNewVideo
          type: boolean
        autoCreatePrediction:
          description: Automatically create predictions for new videos
          example: false
          title: autoCreatePrediction
          type: boolean
      title: ChannelSettings
      type: object
...
    Pagination:
      example:
        totalItems: 127
        hasNextPage: true
        limit: 20
        totalPages: 7
        hasPreviousPage: false
        page: 1
      properties:
        page:
          example: 1
          title: page
          type: integer
        limit:
          example: 20
          title: limit
          type: integer
        totalItems:
          example: 127
          title: totalItems
          type: integer
        totalPages:
          example: 7
          title: totalPages
          type: integer
        hasNextPage:
          example: true
          title: hasNextPage
          type: boolean
        hasPreviousPage:
          example: false
          title: hasPreviousPage
          type: boolean
      title: Pagination
      type: object
...
    _channels_callback_get_200_response:
      example:
        data:
          videoCount: 342
          thumbnail: https://yt3.ggpht.com/abc123
          connectedAt: 2025-09-01T10:00:00Z
          channelUrl: https://youtube.com/@techtutorialshub
          lastSyncAt: 2025-10-03T08:00:00Z
          subscriberCount: 250000
          channelName: Tech Tutorials Hub
          viewCount: 15000000
          channelId: UCabc123xyz
          syncStatus: active
        success: true
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          $ref: "#/components/schemas/ConnectedChannel"
      title: _channels_callback_get_200_response
      type: object
    _channels_me_get_404_response_error:
      example:
        code: NO_CHANNEL_CONNECTED
        message: No YouTube channel is connected to this account
      properties:
        code:
          example: NO_CHANNEL_CONNECTED
          title: code
          type: string
        message:
          example: No YouTube channel is connected to this account
          title: message
          type: string
      title: _channels_me_get_404_response_error
      type: object
    _channels_me_get_404_response:
      example:
        success: false
        error:
          code: NO_CHANNEL_CONNECTED
          message: No YouTube channel is connected to this account
      properties:
        success:
          example: false
          title: success
          type: boolean
        error:
          $ref: "#/components/schemas/_channels_me_get_404_response_error"
      title: _channels_me_get_404_response
      type: object
    _channels_me_delete_200_response:
      example:
        success: true
        message: Channel disconnected successfully
      properties:
        success:
          example: true
          title: success
          type: boolean
        message:
          example: Channel disconnected successfully
          title: message
          type: string
      title: _channels_me_delete_200_response
      type: object
...
    _channels_me_videos_get_200_response:
      example:
        pagination:
          totalItems: 127
          hasNextPage: true
          limit: 20
          totalPages: 7
          hasPreviousPage: false
          page: 1
        data:
        - duration: 720
          thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg
          publishedAt: 2025-09-15T14:00:00Z
          hasPrediction: true
          description: Learn something amazing in this video
          videoId: dQw4w9WgXcQ
          likeCount: 8500
          viewCount: 125000
          title: Amazing Tutorial Video
          predictionId: pred_abc123
          commentCount: 342
        - duration: 720
          thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/maxresdefault.jpg
          publishedAt: 2025-09-15T14:00:00Z
          hasPrediction: true
          description: Learn something amazing in this video
          videoId: dQw4w9WgXcQ
          likeCount: 8500
          viewCount: 125000
          title: Amazing Tutorial Video
          predictionId: pred_abc123
          commentCount: 342
        success: true
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          items:
            $ref: "#/components/schemas/ChannelVideo"
          title: data
          type: array
        pagination:
          $ref: "#/components/schemas/Pagination"
      title: _channels_me_videos_get_200_response
      type: object
    _channels_me_videos_get_404_response_error:
      example:
        code: NO_CHANNEL_CONNECTED
        message: No YouTube channel is connected
      properties:
        code:
          example: NO_CHANNEL_CONNECTED
          title: code
          type: string
        message:
          example: No YouTube channel is connected
          title: message
          type: string
      title: _channels_me_videos_get_404_response_error
      type: object
    _channels_me_videos_get_404_response:
      example:
        success: false
        error:
          code: NO_CHANNEL_CONNECTED
          message: No YouTube channel is connected
      properties:
        success:
          example: false
          title: success
          type: boolean
        error:
          $ref: "#/components/schemas/_channels_me_videos_get_404_response_error"
      title: _channels_me_videos_get_404_response
      type: object
    _channels_me_analytics_get_200_response:
      example:
        data:
          period: 30d
          timeline:
          - date: 2025-10-01
            subscribers: 85
            estimatedMinutesWatched: 3500
            views: 15000
          - date: 2025-10-01
            subscribers: 85
            estimatedMinutesWatched: 3500
            views: 15000
          metrics:
            subscribers: 2500
            estimatedMinutesWatched: 125000
            averageViewDuration: 180
            views: 500000
          topVideos:
          - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
            videoId: dQw4w9WgXcQ
            title: Top Performing Video
            views: 50000
          - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
            videoId: dQw4w9WgXcQ
            title: Top Performing Video
            views: 50000
          - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
            videoId: dQw4w9WgXcQ
            title: Top Performing Video
            views: 50000
          - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
            videoId: dQw4w9WgXcQ
            title: Top Performing Video
            views: 50000
          - thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
            videoId: dQw4w9WgXcQ
            title: Top Performing Video
            views: 50000
          demographics:
            gender:
              other: 1.7
              female: 35.8
              male: 62.5
            topCountries:
            - country: United States
              countryCode: US
              percentage: 45.2
            - country: United States
              countryCode: US
              percentage: 45.2
            ageGroups:
            - ageRange: 18-24
              percentage: 35.5
            - ageRange: 18-24
              percentage: 35.5
        success: true
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          $ref: "#/components/schemas/ChannelAnalytics"
      title: _channels_me_analytics_get_200_response
      type: object
    _channels_me_settings_get_200_response:
      example:
        data:
          syncFrequency: daily
          autoCreatePrediction: false
          autoTracking: true
          notifyOnNewVideo: true
        success: true
      properties:
        success:
          example: true
          title: success
          type: boolean
        data:
          $ref: "#/components/schemas/ChannelSettings"
      title: _channels_me_settings_get_200_response
      type: object
...
    ChannelAnalytics_metrics:
      example:
        subscribers: 2500
        estimatedMinutesWatched: 125000
        averageViewDuration: 180
        views: 500000
      properties:
        views:
          example: 500000
          title: views
          type: integer
        subscribers:
          example: 2500
          title: subscribers
          type: integer
        estimatedMinutesWatched:
          example: 125000
          title: estimatedMinutesWatched
          type: integer
        averageViewDuration:
          description: Average view duration in seconds
          example: 180
          title: averageViewDuration
          type: integer
      title: ChannelAnalytics_metrics
      type: object
    ChannelAnalytics_timeline_inner:
      example:
        date: 2025-10-01
        subscribers: 85
        estimatedMinutesWatched: 3500
        views: 15000
      properties:
        date:
          example: 2025-10-01
          format: date
          title: date
          type: string
        views:
          example: 15000
          title: views
          type: integer
        subscribers:
          example: 85
          title: subscribers
          type: integer
        estimatedMinutesWatched:
          example: 3500
          title: estimatedMinutesWatched
          type: integer
      title: ChannelAnalytics_timeline_inner
      type: object
    ChannelAnalytics_topVideos_inner:
      example:
        thumbnail: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
        videoId: dQw4w9WgXcQ
        title: Top Performing Video
        views: 50000
      properties:
        videoId:
          example: dQw4w9WgXcQ
          title: videoId
          type: string
        title:
          example: Top Performing Video
          title: title
          type: string
        views:
          example: 50000
          title: views
          type: integer
        thumbnail:
          example: https://i.ytimg.com/vi/dQw4w9WgXcQ/default.jpg
          format: uri
          title: thumbnail
          type: string
      title: ChannelAnalytics_topVideos_inner
      type: object
    ChannelAnalytics_demographics_ageGroups_inner:
      example:
        ageRange: 18-24
        percentage: 35.5
      properties:
        ageRange:
          example: 18-24
          title: ageRange
          type: string
        percentage:
          example: 35.5
          format: float
          title: percentage
          type: number
      title: ChannelAnalytics_demographics_ageGroups_inner
      type: object
    ChannelAnalytics_demographics_gender:
      example:
        other: 1.7
        female: 35.8
        male: 62.5
      properties:
        male:
          example: 62.5
          format: float
          title: male
          type: number
        female:
          example: 35.8
          format: float
          title: female
          type: number
        other:
          example: 1.7
          format: float
          title: other
          type: number
      title: ChannelAnalytics_demographics_gender
      type: object
    ChannelAnalytics_demographics_topCountries_inner:
      example:
        country: United States
        countryCode: US
        percentage: 45.2
      properties:
        country:
          example: United States
          title: country
          type: string
        countryCode:
          example: US
          title: countryCode
          type: string
        percentage:
          example: 45.2
          format: float
          title: percentage
          type: number
      title: ChannelAnalytics_demographics_topCountries_inner
      type: object
    ChannelAnalytics_demographics:
      example:
        gender:
          other: 1.7
          female: 35.8
          male: 62.5
        topCountries:
        - country: United States
          countryCode: US
          percentage: 45.2
        - country: United States
          countryCode: US
          percentage: 45.2
        ageGroups:
        - ageRange: 18-24
          percentage: 35.5
        - ageRange: 18-24
          percentage: 35.5
      properties:
        ageGroups:
          items:
            $ref: "#/components/schemas/ChannelAnalytics_demographics_ageGroups_inner"
          title: ageGroups
          type: array
        gender:
          $ref: "#/components/schemas/ChannelAnalytics_demographics_gender"
        topCountries:
          items:
            $ref: "#/components/schemas/ChannelAnalytics_demographics_topCountries_inner"
          title: topCountries
          type: array
      title: ChannelAnalytics_demographics
      type: object
    inline_object_error_details_inner:
      example:
        field: email
        message: Invalid email format
      properties:
        field:
          example: email
          title: field
          type: string
        message:
          example: Invalid email format
          title: message
          type: string
      title: inline_object_error_details_inner
      type: object
    inline_object_error:
      example:
        code: VALIDATION_ERROR
        details:
        - field: email
          message: Invalid email format
        - field: email
          message: Invalid email format
        message: Validation failed
      properties:
        code:
          example: VALIDATION_ERROR
          title: code
          type: string
        message:
          example: Validation failed
          title: message
          type: string
        details:
          items:
            $ref: "#/components/schemas/inline_object_error_details_inner"
          title: details
          type: array
      title: inline_object_error
      type: object
    inline_object:
      example:
        success: false
        error:
          code: VALIDATION_ERROR
          details:
          - field: email
            message: Invalid email format
          - field: email
            message: Invalid email format
          message: Validation failed
      properties:
        success:
          example: false
          title: success
          type: boolean
        error:
          $ref: "#/components/schemas/inline_object_error"
      title: inline_object
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: JWT
      description: JWT token obtained from login or registration
      scheme: bearer
      type: http
...